requirements.txt
beautifulsoup4==4.12.2
openai==0.28.1
python-dotenv==1.0.0
requests==2.31.0
